remove renderers.py and test_renderers.py altogether and migrate remaining DomeRenderer to operators
