+++ pyproject.toml
@@ -39,7 +39,6 @@ target-version = "py38"
 "tests/test_loaders.py" = ["N802", "N803"]
 "tests/test_dataclass.py" = ["F811"]
 "src/unitxt/validate.py" = ["B024"]
-"src/unitxt/renderers.py" = ["B024"]
 "src/unitxt/type_utils.py" = ["C901"]
 "src/unitxt/dataclass.py" = ["C901"]
 "src/unitxt/operators.py" = ["C901"]
