+++ src/unitxt/dataset.py
@@ -29,7 +29,6 @@ from .random_utils import __file__ as _
 from .recipe import __file__ as _
 from .register import __file__ as _
 from .register import _reset_env_local_catalogs, register_all_artifacts
-from .renderers import __file__ as _
 from .schema import __file__ as _
 from .split_utils import __file__ as _
 from .splitters import __file__ as _
